Hand Gestures with AI/requirements.txt
opencv-python
pillow
cvzone
mediapipe
numpy
google-generativeai
streamlit
